+++ skill-share/src/components/navbar/NavBar.css
@@ -1,14 +1,12 @@
 .navbar {
     height: 95vh;
-    width: 20%;
+    width: 100%;
     display: flex;
     flex-direction: column;
     justify-content: space-between;
     top: 0px;
     left: 0px;
-    position: fixed;
-    z-index: 1;
-    margin-left: 2%;
+    position: relative;
 
     background-color: white;
 }
