+++ skill-share/src/pages/explore-page/Explore.css
@@ -1,11 +1,10 @@
-* {
+/* * {
   box-sizing: border-box;
   margin: 0;
   padding: 0;
-}
+} */
 
 .explore-main {
-  margin-left: 220px; /* width of navigation bar */
   padding: 2rem;
   min-height: 100vh;
   width: 100% ;      /*updated */
