+++ skill-share/src/pages/profile-page/Profile.css
@@ -3,17 +3,18 @@
     flex-direction: row;
     position: relative;
     background-color: #fafafa;
-    left: 20%;
-    width: 75%;
-    max-width: 75%;
+    left: 0;
+    width: 95%;
+    max-width: 95%;
 }
 
 .profile {
     flex: 1;
     min-height: 100vh;
-    padding: 40px 80px;
+    padding: 40px 0px;
     background-color: white;
     color: #262626;
+    max-width: 100%;
 }
 
 /* Edit Profile Modal */
@@ -342,11 +343,12 @@
     margin-top: 20px;
 }
 
-.posts-grid {
+.profile-posts-grid {
     display: grid;
     place-items: center;
     grid-template-columns: repeat(3, 1fr);
     gap: 4px;
+    max-width: 100%;
 }
 
 .post-item {
